backend/pyproject.toml
[project]
name = "pissgraph-backend"
version = "0.1.0"
description = "ISS urine tank telemetry backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "lightstreamer-client-lib>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncio-mqtt>=0.16.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pissgraph"]

[tool.uv]
dev-dependencies = [
    "mypy>=1.7.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]